Undo switched names
